variable collector
